refactor(didcomm): collapse v2 forward builder into DidCommForwardMessageV2 static
