+++ packages/didcomm/src/modules/routing/messages/v2/index.ts
@@ -1,4 +1,3 @@
-export * from './ForwardMessage'
 export * from './ForwardMessageV2'
 export * from './KeylistMessage'
 export * from './KeylistQueryMessage'
